refactor: Migrate Metals server to Scala 3
This mostly involved:
- changing imports
- moving things between MtagsEnrichements variants
- adding ()
- fixing bugs reported now as warning or errors (things previously not reported at all)

+++ mtags/src/main/scala/scala/meta/internal/metals/URIEncoderDecoder.scala
@@ -6,7 +6,7 @@ object URIEncoderDecoder {
   // Currently Metals doesn't encode `:` but does decode it
   private val toEscape: Map[Char, String] =
     Set('"', '<', '>', '&', '\'', '[', ']', '{', '}', ' ', '+', '!')
-      .map(char => char -> ("%" + char.toHexString))
+      .map(char => char -> ("%" + char.toInt.toHexString))
       .toMap
 
   private val toDecode: Map[String, Char] = toEscape.map { case (k, v) =>
